backend/update-work-preview/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
rarfile==4.1